pyproject.toml
[project]
name = "gh-issue-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.6.0",
    "dotenv>=0.9.9",
    "langchain-core>=0.3.68",
    "litellm[caching]>=1.74.0.post1",
    "pydantic>=2.11.7",
    "richer>=0.1.6",
    "tqdm>=4.67.1",
    "weave>=0.51.56",
]
